+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * New environment variable `PINS_CACHE_DIR` controls the location of the 
   default cache path (#748).
+  
+* Added new board for Google Drive `board_gdrive()` (#749).
 
 # pins 1.2.0